Add files via upload
Task 1. Write a script that will parse data from the attached csv file, then create and insert the csv data into a table in a mysql database. 
            Please use database connection info is provided below.
Task 2. Write a 2nd script that will query the table from task 1, and calculate and display the average of all values from column "temp".

Note: Because here is a public access repository, the username/password are not included.